hibernate update

+++ core/src/main/java/com/epam/esm/service/template/OrderService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.epam.esm.repository.model.Order;
 
 public interface OrderService {
-    List<Order> getAll(long limit,long offset);
+    List<Order> getAll(int limit,int offset);
     Order getById(long orderId);
     Order update(Order orderPatch,long orderId);
     void delete(long orderId);
