+++ core/src/main/java/com/epam/esm/service/template/TagService.java
@@ -5,7 +5,7 @@ import com.epam.esm.repository.model.Tag;
 import java.util.List;
 
 public interface TagService {
-    List<Tag> getAll(long limit,long offset);
+    List<Tag> getAll(int limit,int offset);
     Tag getByID(long id);
     Tag addEntity(Tag tag);
     void deleteByID(long id);
