+++ core/src/main/java/com/epam/esm/service/template/UserService.java
@@ -6,7 +6,7 @@ import com.epam.esm.repository.model.Tag;
 import com.epam.esm.repository.model.User;
 
 public interface UserService {
-     List<User> getAll(long limit,long offset);
+     List<User> getAll(int limit,int offset);
      User getById(long id);
      User getByName(String userName);
      Tag fetchMostUsedTagWithRichestOrders();
